backend_fastapi/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymongo[srv]==4.6.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]
mkdocs>=1.5.0
mkdocs-material>=9.0.0==3.3.0
